[#7]fix: 수정된 Order에 맞춰 payment 로직 수정

+++ src/main/java/org/example/mollyapi/payment/service/PaymentService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface PaymentService {
 
     //결제 승인 절차
-    public Payment processPayment(Long userId, String paymentKey, String tossOrderId, Integer amount, Integer point, Long deliveryId, String paymentType);
+    public Payment processPayment(Long userId, String paymentKey, String tossOrderId, Long amount, Integer point, Long deliveryId, String paymentType);
 
     //결제 성공 절차
     public void successPayment(Payment payment, String tossOrderId,Integer point);
